refactor: Participant port 패키지 변경 및 네이밍 변경
